feat: redis 이용한 장바구니 구현

+++ src/main/java/com/nhnacademy/bookstoreback/bookcart/domain/dto/request/UpdateBookCartRequest.java
@@ -1,4 +1,4 @@
 package com.nhnacademy.bookstoreback.bookcart.domain.dto.request;
 
-public record UpdateBookCartRequest(int bookQuantity) {
+public record UpdateBookCartRequest(Long bookQuantity) {
 }
